+++ src/controllers/CampaignTypesController.php
@@ -88,7 +88,6 @@ class CampaignTypesController extends Controller
         $campaignType->siteId = $this->request->getBodyParam('siteId', $campaignType->siteId);
         $campaignType->name = $this->request->getBodyParam('name', $campaignType->name);
         $campaignType->handle = $this->request->getBodyParam('handle', $campaignType->handle);
-        $campaignType->enableAnonymousTracking = $this->request->getBodyParam('enableAnonymousTracking', $campaignType->enableAnonymousTracking);
         $campaignType->uriFormat = $this->request->getBodyParam('uriFormat', $campaignType->uriFormat);
         $campaignType->htmlTemplate = $this->request->getBodyParam('htmlTemplate', $campaignType->htmlTemplate);
         $campaignType->plaintextTemplate = $this->request->getBodyParam('plaintextTemplate', $campaignType->plaintextTemplate);