+++ src/services/CampaignsService.php
@@ -106,8 +106,8 @@ class CampaignsService extends Component
             }
         }
 
-        // Only save is anonymous tracking is not enabled
-        if (!$campaignRecord->campaignType->enableAnonymousTracking) {
+        // Only save if anonymous tracking is not enabled
+        if (!Campaign::$plugin->getSettings()->enableAnonymousTracking) {
             $contactCampaignRecord->save();
         }