+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.0-beta.2 - 2022-04-17
 ### Added
 - Added an "Export to CSV" button to all datatables in reports ([#245](https://github.com/putyourlightson/craft-campaign/issues/245)).
+- Added the `enableAnonymousTracking` setting, which prevents tracking of contact interactions ([#115](https://github.com/putyourlightson/craft-campaign/issues/115)).
 
 ### Fixed
 - Fixed a bug in which the "New" button was not appearing in element type modal windows.